+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Register the Options update hook only in the admin/backoffice.
+
 ## [0.5.8] - 2025-01-13
 
 ### Added