+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.9] - 2025-01-13
+
 ### Changed
 
 - Register the Options update hook only in the admin/backoffice.
@@ -549,7 +551,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.5.8...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.5.9...HEAD
+[0.5.9]: https://github.com/26b/unbabble/compare/0.5.8...0.5.9
 [0.5.8]: https://github.com/26b/unbabble/compare/0.5.7...0.5.8
 [0.5.7]: https://github.com/26b/unbabble/compare/0.5.6...0.5.7
 [0.5.6]: https://github.com/26b/unbabble/compare/0.5.5...0.5.6